refactor: std/pairing have more consistent apis
