+++ std/algebra/fields_bls24315/e24.go
@@ -467,54 +467,3 @@ func (e *E24) FrobeniusQuad(api frontend.API, x E24) *E24 {
 
 	return e
 }
-
-// FinalExponentiation computes the final expo x**(p**12-1)(p**4+1)(p**8 - p**4 +1)/r
-func (e *E24) FinalExponentiation(api frontend.API, e1 E24) *E24 {
-	const ateLoop = 3218079743
-	const genT = ateLoop
-	result := e1
-
-	// https://eprint.iacr.org/2012/232.pdf, section 7
-	var t [9]E24
-
-	// easy part
-	t[0].Conjugate(api, result)
-	result.Inverse(api, result)
-	t[0].Mul(api, t[0], result)
-	result.FrobeniusQuad(api, t[0]).
-		Mul(api, result, t[0])
-
-	// hard part (api, up to permutation)
-	// Daiki Hayashida and Kenichiro Hayasaka
-	// and Tadanori Teruya
-	// https://eprint.iacr.org/2020/875.pdf
-	// 3*Phi_24(api, p)/r = (api, u-1)² * (api, u+p) * (api, u²+p²) * (api, u⁴+p⁴-1) + 3
-	t[0].CyclotomicSquare(api, result)
-	t[1].Expt(api, result, genT)
-	t[2].Conjugate(api, result)
-	t[1].Mul(api, t[1], t[2])
-	t[2].Expt(api, t[1], genT)
-	t[1].Conjugate(api, t[1])
-	t[1].Mul(api, t[1], t[2])
-	t[2].Expt(api, t[1], genT)
-	t[1].Frobenius(api, t[1])
-	t[1].Mul(api, t[1], t[2])
-	result.Mul(api, result, t[0])
-	t[0].Expt(api, t[1], genT)
-	t[2].Expt(api, t[0], genT)
-	t[0].FrobeniusSquare(api, t[1])
-	t[2].Mul(api, t[0], t[2])
-	t[1].Expt(api, t[2], genT)
-	t[1].Expt(api, t[1], genT)
-	t[1].Expt(api, t[1], genT)
-	t[1].Expt(api, t[1], genT)
-	t[0].FrobeniusQuad(api, t[2])
-	t[0].Mul(api, t[0], t[1])
-	t[2].Conjugate(api, t[2])
-	t[0].Mul(api, t[0], t[2])
-	result.Mul(api, result, t[0])
-
-	*e = result
-
-	return e
-}
